feat(vmux_agent): wire PageAgentPlugin into desktop, supersede AgentPage
- Replace vmux_agent::AgentPage with vmux_agent::PageAgentPlugin in
  vmux_desktop/lib.rs. PageAgentPlugin is a strict superset: includes
  continue_after_tool + surface_errors + AgentToast emitter.
- Fold settings-based echo strategy registration into PageAgentPlugin
  as a Startup system (after SettingsLoadSet).
- Add register_builtin_providers Startup system: for each builtin
  (mistral/anthropic/openai), if its env var is set, register the
  matching AgentPageStrategy with its default model.
- Add AgentStrategies::register_page_if_absent so user-supplied
  strategies in tests are not overwritten by startup registration.
- Delete client/page/agent.rs and the AgentPage re-export.
