+++ crates/vmux_agent/src/client/page.rs
@@ -1,3 +1,2 @@
-pub mod agent;
 pub mod plugin;
 pub mod strategy;
