+++ crates/vmux_agent/src/client/page/plugin.rs
@@ -1,7 +1,12 @@
+use std::sync::Arc;
+
 use bevy::prelude::*;
 use bevy_cef::prelude::BinEventEmitterPlugin;
+use vmux_setting::{AppSettings, SettingsLoadSet};
 
 use crate::components::{AgentApprovalPolicy, AgentMessages, AgentSession};
+use crate::echo::EchoPageStrategy;
+use crate::providers::{BUILTIN_PROVIDERS, instantiate_builtin};
 use crate::run_state_kind::LastRunStateKind;
 use crate::strategy::AgentStrategies;
 use crate::systems::{
@@ -31,11 +36,62 @@ impl Plugin for PageAgentPlugin {
                     surface_errors::surface_errors,
                     attach_last_run_state_kind,
                 ),
+            )
+            .add_systems(
+                Startup,
+                (
+                    register_page_agents_from_settings,
+                    register_builtin_providers,
+                )
+                    .after(SettingsLoadSet),
             );
 
         if app.world().get_resource::<AgentStrategies>().is_none() {
             app.insert_resource(AgentStrategies::default());
         }
+    }
+}
+
+fn register_page_agents_from_settings(
+    settings: Option<Res<AppSettings>>,
+    strategies: Option<ResMut<AgentStrategies>>,
+) {
+    let Some(settings) = settings else { return };
+    let Some(mut strategies) = strategies else {
+        return;
+    };
+    for provider_settings in &settings.agent.app_providers {
+        let kind = match provider_settings.kind.as_str() {
+            "vibe" => vmux_core::agent::AgentKind::Vibe,
+            "claude" => vmux_core::agent::AgentKind::Claude,
+            "codex" => vmux_core::agent::AgentKind::Codex,
+            other => {
+                bevy::log::warn!(
+                    "agent.app_providers: unknown kind '{other}' for provider '{}'; defaulting to vibe",
+                    provider_settings.provider
+                );
+                vmux_core::agent::AgentKind::Vibe
+            }
+        };
+        for model in &provider_settings.models {
+            strategies.register_page_if_absent(Arc::new(EchoPageStrategy::new(
+                provider_settings.provider.clone(),
+                model.clone(),
+                kind,
+            )));
+        }
+    }
+}
+
+fn register_builtin_providers(strategies: Option<ResMut<AgentStrategies>>) {
+    let Some(mut strategies) = strategies else {
+        return;
+    };
+    for p in BUILTIN_PROVIDERS {
+        if std::env::var(p.env_var).is_err() {
+            continue;
+        }
+        strategies.register_page_if_absent(instantiate_builtin(p, p.default_model));
     }
 }
